Lab3Test

+++ vscode/Chapter03/CheckingForOverflow/Program.cs
@@ -2,7 +2,7 @@
 {
   checked
   {
-    int x = int.MaxValue - 1;
+    double x = double.MaxValue - 1;
     WriteLine($"Initial value: {x}");
     x++;
     WriteLine($"After incrementing: {x}");
@@ -19,7 +19,7 @@ catch (OverflowException)
 
 unchecked
 {
-  int y = int.MaxValue + 1;
+  double y = double.MaxValue + 1;
   WriteLine($"Initial value: {y}");
   y--;
   WriteLine($"After decrementing: {y}");
